refactor: move hook into separate file
organize hooks into separate file inside the /hooks folder
