+++ app/components/Base/RemoteImage/index.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import RemoteImage, { useSVGViewBox } from './';
-import { renderHook, waitFor } from '@testing-library/react-native';
+import RemoteImage from './';
 
 jest.mock('react-redux', () => ({
   ...jest.requireActual('react-redux'),
@@ -44,45 +43,5 @@ describe('RemoteImage', () => {
       />,
     );
     expect(wrapper).toMatchSnapshot();
-  });
-});
-
-describe('useSVGViewBox()', () => {
-  const MOCK_SVG_WITH_VIEWBOX = `<svg xmlns="http://www.w3.org/2000/svg" width="60" height="60" viewBox="0 0 60 60" fill="none"></svg>`;
-  const MOCK_SVG_WITHOUT_VIEWBOX = `<svg xmlns="http://www.w3.org/2000/svg" width="60" height="60" fill="none"></svg>`;
-
-  function arrangeMocks() {
-    const mockResponseTextFn = jest
-      .fn()
-      .mockResolvedValue(MOCK_SVG_WITHOUT_VIEWBOX);
-    jest
-      .spyOn(global, 'fetch')
-      .mockResolvedValue({ text: mockResponseTextFn } as unknown as Response);
-
-    return { mockText: mockResponseTextFn };
-  }
-
-  it('should return view-box if svg if missing a view-box', async () => {
-    const { mockText } = arrangeMocks();
-    mockText.mockResolvedValueOnce(MOCK_SVG_WITHOUT_VIEWBOX);
-
-    const hook = renderHook(() => useSVGViewBox('URI', true));
-    await waitFor(() => expect(hook.result.current).toBeDefined());
-  });
-
-  it('should not return a view-box if svg already has view-box', async () => {
-    const { mockText } = arrangeMocks();
-    mockText.mockResolvedValueOnce(MOCK_SVG_WITH_VIEWBOX);
-
-    const hook = renderHook(() => useSVGViewBox('URI', true));
-    await waitFor(() => expect(hook.result.current).toBeUndefined());
-  });
-
-  it('should not make async calls if image is not an svg', async () => {
-    const mocks = arrangeMocks();
-    const hook = renderHook(() => useSVGViewBox('URI', false));
-
-    await waitFor(() => expect(hook.result.current).toBeUndefined());
-    expect(mocks.mockText).not.toHaveBeenCalled();
   });
 });
